+++ provider/pkg/provider/ids_test.go
@@ -15,11 +15,12 @@
 package provider
 
 import (
+	"strings"
+	"testing"
+
 	"github.com/pulumi/pulumi-google-native/provider/pkg/resources"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
 	"github.com/stretchr/testify/assert"
-	"strings"
-	"testing"
 )
 
 func TestCalculateResourceId_IdProperty(t *testing.T) {
@@ -70,8 +71,9 @@ func TestGetDefaultName_Generated(t *testing.T) {
 		"location": "west",
 		"ignoreMe": 1.2,
 	})
-	actual := getDefaultName(urn, "projects/{project}/locations/{location}/things/{name}", olds, news)
+	actual, autonamed := getDefaultName(urn, "projects/{project}/locations/{location}/things/{name}", olds, news)
 	expectedPrefix := "projects/p01/locations/west/things/myName-"
+	assert.True(t, autonamed)
 	assert.True(t, strings.HasPrefix(actual.StringValue(), expectedPrefix))
 	assert.Equal(t, len(expectedPrefix)+7, len(actual.StringValue()))
 }
@@ -87,7 +89,8 @@ func TestGetDefaultName_OldApplied(t *testing.T) {
 		"location": "west",
 		"ignoreMe": 1.2,
 	})
-	actual := getDefaultName(urn, "projects/{project}/locations/{location}/things/{name}", olds, news)
+	actual, autonamed := getDefaultName(urn, "projects/{project}/locations/{location}/things/{name}", olds, news)
+	assert.False(t, autonamed)
 	assert.Equal(t, fixedName, actual.StringValue())
 }